rename main FC ProductItem => Education, rearenge structure in one folder education
