--- education/src/blocks/product/ProductClient.styled.ts
@@ -1,5 +1,0 @@
-import styled from 'styled-components';
-
-export const ProductClientWrapper = styled.div`
-
-`;
