+++ education/src/education/blocks/product/Education.styled.ts
@@ -1,0 +1,5 @@
+import styled from 'styled-components';
+
+export const EducationWrapper = styled.div`
+
+`;
