+++ education/src/education/components/ProductClientSections/ProductItemRecord.tsx
@@ -1,6 +1,6 @@
 import React, { FC } from "react"
-import { Display } from "../../ui/styled/General.styled";
-import { TitleWrapper, BtnWrapper } from "../../ui/styled/ProductItem.styled";
+import { Display } from "../../styled/General.styled";
+import { TitleWrapper, BtnWrapper } from "../../styled/ProductItem.styled";
 import ProductItem from "../ProductItem/ProductItem";
 
 interface IProductItemRecord{
