+++ education/src/education/components/ProductClientSections/ProductOrderRecord.tsx
@@ -1,10 +1,10 @@
 import { FC, useState, useEffect } from "react";
 import axios from 'axios';
-import { Display } from "../../ui/styled/General.styled";
-import { TitleWrapper, BtnWrapper } from "../../ui/styled/ProductItem.styled";
+import { Display } from "../../styled/General.styled";
+import { TitleWrapper, BtnWrapper } from "../../styled/ProductItem.styled";
 import ProductOrder from "../ProductItem/ProductOrder";
 import { dictNameToRoute } from "../../contexts/ProductItemData";
-import { TClientProperty } from '../../blocks/product/ProductClient';
+import { TClientProperty } from '../../blocks/product/Education';
 
 interface IProductOrderRecord{
     _width?: number,
@@ -35,17 +35,16 @@ const ProductOrderRecord: FC<IProductOrderRecord> = (props) => {
             }
         });
     
+        setResponce('');
+        
         let queryName = dictNameToRoute[props.nameProduct == null ? 'none' : props.nameProduct] == undefined ? props.nameProduct : dictNameToRoute[props.nameProduct == null ? 'none' : props.nameProduct];
         productOrderPost.post(`product-order-dto?name=${queryName}`,  props._clientProperty)
         .then((responce) => {           
             setProductOrder(responce.data);
         })
         .catch((error) => {
-            console.log(error);
             setProductOrder(null);
         });       
-
-        console.log('props._clientProperty');
     }
 
     useEffect(handleProductOrder, [props.nameProduct, props._clientProperty]);
