+++ education/src/education/components/ProductItem/ProductItem.tsx
