+++ education/src/education/components/ProductItemDto/TabProductItemDto.tsx
