+++ education/src/education/components/footer/FooterStyles.tsx
