+++ education/src/education/components/options/OptionsEducation.tsx
@@ -1,17 +1,16 @@
 import React, { FC, useState } from "react";
-import { TypeOption, NameOption } from "../../ui/styled/OptionsProductItem.styled";
-import { OptionsEducationWrap } from "../../ui/styled/OptionsEducation.styled";
+import { TypeOption, NameOption } from "../../styled/OptionsProductItem.styled";
+import { OptionsEducationWrap } from "../../styled/OptionsEducation.styled";
 
-
-type TInvitedPerson = {
+interface IOptionsEducation {
     mainName: string,
     optionNames: Array<string>,
-    _handleIsInvited: (e: React.FormEvent<HTMLElement>) => void,
-}
+    _handleEducationForm: (e: React.FormEvent<HTMLElement>) => void,
+};
 
-const OptionsInvited: FC<TInvitedPerson> = (props) => {
+const OptionsEducation: FC<IOptionsEducation> = (props) => {
 
-    const [isActivities, setIsActivities] = useState<Array<boolean>>([true, false])
+    const [isActivities, setIsActivities] = useState<Array<boolean>>([true, false, false])
 
     const handleIsActivities = (index: number, e: React.FormEvent<HTMLElement>):void => {
         const nextIsActivities = isActivities.map((c, i)=> {
@@ -24,16 +23,17 @@ const OptionsInvited: FC<TInvitedPerson> = (props) => {
         });
 
         setIsActivities(nextIsActivities);
-        props._handleIsInvited(e);
+        props._handleEducationForm(e);
     }
 
     return(
         <OptionsEducationWrap>
-            <TypeOption style={{width: '240px'}}>{props.mainName}</TypeOption>
+            <TypeOption>{props.mainName}</TypeOption>
             <NameOption title={props.optionNames[0]} isActive={isActivities[0]} onClick={(e) => {handleIsActivities(0, e)}}>{props.optionNames[0].toUpperCase()}</NameOption>
             <NameOption title={props.optionNames[1]} isActive={isActivities[1]} onClick={(e) => {handleIsActivities(1, e)}}>{props.optionNames[1].toUpperCase()}</NameOption>
+            <NameOption title={props.optionNames[2]} isActive={isActivities[2]} onClick={(e) => {handleIsActivities(2, e)}}>{props.optionNames[2].toUpperCase()}</NameOption>
         </OptionsEducationWrap>
     );
-}
+};
 
-export default OptionsInvited;
+export default OptionsEducation;
