+++ education/src/education/components/styles/TableProduct.styled.ts
