+++ education/src/education/styled/General.styled.ts
