+++ education/src/education/styled/OptionsEducation.styled.ts
